feat: add feature and fix volume bugs

+++ src/Commands/Setup.cs
@@ -21,45 +21,57 @@ namespace Plugin.Tezos.Commands
 
         private static string secretWallet = "./db.json";
 
-        [KeepixPluginFn("install")]
-        public static async Task<bool> OnInstall(InstallInput input)
+        [KeepixPluginFn("install")]     
+        public static async Task<bool> OnInstall(WalletInput input)
         {
+
+            var allRulesPassed = await SetupService.ApplyRules(SetupService.IsDockerRunning, SetupService.IsNotContainnerRunning, SetupService.IsCliInstalled);
+            if (!allRulesPassed)
+            {
+                return false;
+            }
 
             dynamic metadata = JObject.Parse(await ProcessService.ExecuteCommand("curl", $" -L  {Configurations.TEZOS_SNAPSHOT}"));
             if (metadata.url == null || metadata.url == "")
-                return false;
-            using (var client = new HttpClient())
             {
-
-                Console.WriteLine("Start downloading");
-                try
-                {
-                    await client.DownloadFileAsync((string)metadata.url, "/root/tezos-mainnet.rolling");
-                }
-                catch (Exception ex)
-                {
-                    // log error
-                    Console.WriteLine(ex);
-                }
+                LoggerService.Log("Can't get url of the snapshot");
+                return false;
             }
 
+       /*     var isSnapshotDownloaded =  await SetupService.DownloadSnapshot((string)metadata.url, "/root/tezos-mainnet.rolling");
+            if (!isSnapshotDownloaded)
+            {
+                LoggerService.Log("Downloaded of the snapshot failed");
+                return false;
+            }*/
 
-            Console.WriteLine("Start sync");
             await File.WriteAllTextAsync(secretWallet, input.WalletSecretKey);
 
-            await ProcessService.ExecuteCommand("docker","compose up -d node_rolling");
-            await ProcessService.ExecuteCommand("docker", "exec octez-public-node-rolling rm /var/run/tezos/node/data/lock");
-            await ProcessService.ExecuteCommand("docker", "exec octez-public-node-rolling rm -r /var/run/tezos/node/data");
-            await ProcessService.ExecuteCommand("docker","compose stop node_rolling");
-            await ProcessService.ExecuteCommand("rm", "-rf /var/lib/docker/volumes/mainnet-node/_data/data/context");
-            await ProcessService.ExecuteCommand("rm", "-rf /var/lib/docker/volumes/mainnet-node/_data/data/store");
-            await ProcessService.ExecuteCommand("rm", "-rf /var/lib/docker/volumes/mainnet-node/_data/data/lock");
-            await ProcessService.ExecuteCommand("rm", "-rf /var/lib/docker/volumes/mainnet-node/_data/data/daily_logs/");
-            await ProcessService.ExecuteCommand("docker", "compose up import");
-            await ProcessService.ExecuteCommand("docker", "compose up -d node_rolling");
-            await ProcessService.ExecuteCommand("octez-client", "--endpoint http://localhost:8732 config update");
-            await ProcessService.ExecuteCommand("octez-client", $"import secret key {input.WalletName} {input.WalletSecretKey}");
+            try
+            {
+                LoggerService.Log("Starting to setup the tezos container");
 
+                await ProcessService.ExecuteCommand("docker", "volume rm mainnet-client mainnet-node");
+                await ProcessService.ExecuteCommand("docker", "compose up -d node_rolling");
+                Thread.Sleep(1000);
+                await ProcessService.ExecuteCommand("docker", "exec octez-public-node-rolling rm /var/run/tezos/node/data/lock");
+                await ProcessService.ExecuteCommand("docker", "exec octez-public-node-rolling rm -r /var/run/tezos/node/data");
+                Thread.Sleep(1000);
+                await ProcessService.ExecuteCommand("docker", "compose stop node_rolling"); 
+                await ProcessService.ExecuteCommand("rm", "-rf /var/lib/docker/volumes/mainnet-node/_data/data/context");
+                await ProcessService.ExecuteCommand("rm", "-rf /var/lib/docker/volumes/mainnet-node/_data/data/store");
+                await ProcessService.ExecuteCommand("rm", "-rf /var/lib/docker/volumes/mainnet-node/_data/data/lock");
+                await ProcessService.ExecuteCommand("rm", "-rf /var/lib/docker/volumes/mainnet-node/_data/data/daily_logs/");
+                Thread.Sleep(1000);
+                await ProcessService.ExecuteCommand("docker", "compose up import");
+                await ProcessService.ExecuteCommand("docker", "compose up -d node_rolling");
+                await ProcessService.ExecuteCommand("octez-client", "--endpoint http://localhost:8732 config update");
+                await ProcessService.ExecuteCommand("octez-client", $"import secret key {input.WalletName} {input.WalletSecretKey}");
+            }
+            catch
+            {
+                return false;
+            }
 
             return true;
         }
@@ -67,6 +79,12 @@ namespace Plugin.Tezos.Commands
         [KeepixPluginFn("start")]
         public static async Task<bool> OnStartFunc()
         {
+            var allRulesPassed = await SetupService.ApplyRules(SetupService.IsDockerRunning, SetupService.IsNotContainnerRunning);
+            if (!allRulesPassed)
+            {
+                return false;
+            }
+
             await ProcessService.ExecuteCommand("docker", "compose up -d node_rolling");
     
             return true;
@@ -75,6 +93,13 @@ namespace Plugin.Tezos.Commands
         [KeepixPluginFn("stop")]
         public static async Task<bool> OnStopFunc()
         {
+            var allRulesPassed = await SetupService.ApplyRules(SetupService.IsDockerRunning, SetupService.IsContainnerRunning);
+            if (!allRulesPassed)
+            {
+                return false;
+            }
+
+
             await ProcessService.ExecuteCommand("docker", "stop octez-public-node-rolling");
             return false;
         }
